pyproject.toml
[tool.poetry]
name = "smart-on-fhir-client"
version = "0.0.1"
description = ""
authors = ["Marc <marc@synapse-medicine.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fhirpy = "^1.3.0"
loguru = "^0.5.3"
fastapi = "^0.73.0"
uvicorn = "^0.17.0"
aiocache = "^0.11.1"
PyJWT = "^2.3.0"
cryptography = "^36.0.1"
result = "^0.7.0"
seito = "^0.0.4"
tenacity = "^8.0.1"
"fhir.resources" = "^6.2.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
